Added confirm modal dialog for account creation
