+++ app/components/account-create/template.hbs
@@ -1,0 +1,20 @@
+{{#bs-modal position='center' onHide=(action (optional onCancel)) onSubmit=(action onCreate) as |modal|}}
+  {{#modal.header}}
+    <h4 class="modal-title">
+      {{fa-icon 'exclamation-triangle' class="text-danger"}}
+      {{t 'wallets.accounts.create.title'}}
+    </h4>
+  {{/modal.header}}
+  {{#modal.body}}
+    <p class="lead">{{t 'wallets.accounts.create.lead'}}</p>
+    <p>{{t 'wallets.accounts.create.body' htmlSafe=true}}</p>
+  {{/modal.body}}
+  {{#modal.footer}}
+    {{#bs-button onClick=(action modal.close) type="secondary"}}
+      {{t 'cancel'}}
+    {{/bs-button}}
+    {{#bs-button type="danger" icon="fa fa-plus-square" buttonType='submit'}}
+      {{t 'wallets.accounts.create.createButton'}}
+    {{/bs-button}}
+  {{/modal.footer}}
+{{/bs-modal}}
