+++ app/components/navigation-bar/template.hbs
@@ -22,9 +22,9 @@
         {{/link-to}}
       </li>
       <li class="nav-item">
-        <a href="#" class="nav-link" {{action onCreateAccount wallet}}>
+        {{#link-to 'wallets.overview.accounts.create' wallet class="nav-link"}}
           {{fa-icon "plus-square"}} {{t 'wallets.overview.create'}}
-        </a>
+        {{/link-to}}
       </li>
       <li class="nav-item">
         {{#link-to 'wallets.overview.settings' wallet class="nav-link"}}
