+++ app/router.js
@@ -26,6 +26,7 @@ Router.map(function routerMap() {
     this.route('overview', function overviewRoute() {
       this.route('settings');
       this.route('accounts', { path: '/:account_id' }, function accountsRoute() {
+        this.route('create');
         this.route('send');
         this.route('history');
         this.route('settings');
