+++ app/wallets/overview/accounts/create/template.hbs
@@ -1,0 +1,1 @@
+{{account-create wallet=model onCreate=(action (route-action 'create' model)) onCancel=(action (route-action 'cancel'))}}
