+++ app/wallets/template.hbs
@@ -1,4 +1,4 @@
-{{navigation-bar wallet=model onCreateAccount=(route-action 'createAccount')}}
+{{navigation-bar wallet=model}}
 
 <main role="main">
   {{notification-center}}
