+++ translations/hu-hu.yaml
@@ -126,6 +126,14 @@ wallets:
       none: 'Nincs még számla történet'
     settings:
       representative: 'Képviselő számla'
+    create:
+      title: 'Új számla létrehozása'
+      lead: 'Biztosan szeretnél új számlát létrehozni?'
+      body: >
+        <strong>Figyelmeztetés</strong>: Ez a művelet egy új MIKRON számlaszámot fog létrehozni, amit később nem fogsz tudni
+        törölni. Csak abban az esetben érdemes ezt megtenni, ha a jelenlegi(ek) mellett szeretnél másik MIKRON számlaszámmal
+        is rendelkezni. Biztosan folytatod?
+      createButton: 'Lérehozás'
   overview:
     none: 'Nincs a tárcában számla.'
     create: 'Új számla létrehozása'
